1324 - Understand how interested varies by various fields/Resources/requirements.txt
pandas
mysql-connector-python
dython